chore(OccurrenceTypes): update model, factory, serializer, migration
